acpi::handler

Struct PhysicalMapping

pub struct PhysicalMapping<H, T>
where
    H: AcpiHandler,
{ /* private fields */ }

Describes a physical mapping created by AcpiHandler::map_physical_region and unmapped by AcpiHandler::unmap_physical_region. The region mapped must be at least size_of::<T>() bytes, but may be bigger.

See PhysicalMapping::new for the meaning of each field.

Implementations

impl<H, T> PhysicalMapping<H, T>
where H: AcpiHandler,

pub unsafe fn new( physical_start: usize, virtual_start: NonNull<T>, region_length: usize, mapped_length: usize, handler: H, ) -> Self

Construct a new PhysicalMapping.

physical_start should be the physical address of the structure to be mapped.

virtual_start should be the corresponding virtual address of that structure. It may differ from the start of the region mapped due to requirements of the paging system. It must be a valid, non-null pointer.

region_length should be the number of bytes requested to be mapped. It must be equal to or larger than size_of::<T>().

mapped_length should be the number of bytes mapped to fulfill the request. It may be equal to or larger than region_length, due to requirements of the paging system or other reasoning.

handler should be the same AcpiHandler that created the mapping. When the PhysicalMapping is dropped, it will be used to unmap the structure.

Safety

The caller must ensure that the physical memory can be safely mapped.

pub fn physical_start(&self) -> usize

pub fn virtual_start(&self) -> NonNull<T>

pub fn get(&self) -> Pin<&T>

pub fn region_length(&self) -> usize

pub fn mapped_length(&self) -> usize

pub fn handler(&self) -> &H