acpi::madt

Struct GiccEntry

pub struct GiccEntry {

    pub header: EntryHeader,
    pub cpu_interface_number: u32,
    pub processor_uid: u32,
    pub flags: u32,
    pub parking_protocol_version: u32,
    pub performance_interrupt_gsiv: u32,
    pub parked_address: u64,
    pub gic_registers_address: u64,
    pub gic_virtual_registers_address: u64,
    pub gic_hypervisor_registers_address: u64,
    pub vgic_maintenance_interrupt: u32,
    pub gicr_base_address: u64,
    pub mpidr: u64,
    pub processor_power_efficiency_class: u8,
    pub spe_overflow_interrupt: u16,
    pub trbe_interrupt: ExtendedField<u16, 6>,
    /* private fields */

}

This field will appear for ARM processors that support ACPI and use the Generic Interrupt Controller. In the GICC interrupt model, each logical process has a Processor Device object in the namespace, and uses this structure to convey its GIC information.

Fields
header: EntryHeadercpu_interface_number: u32processor_uid: u32flags: u32parking_protocol_version: u32performance_interrupt_gsiv: u32parked_address: u64gic_registers_address: u64gic_virtual_registers_address: u64gic_hypervisor_registers_address: u64vgic_maintenance_interrupt: u32gicr_base_address: u64mpidr: u64processor_power_efficiency_class: u8spe_overflow_interrupt: u16

SPE overflow Interrupt.

ACPI 6.3 defined this field. It is zero in prior versions or if this processor does not support SPE.

trbe_interrupt: ExtendedField<u16, 6>