acpi

Module mcfg

Structs

Mcfg

McfgEntry

PciConfigEntry

Configuration entry describing a valid bus range for the given PCI segment group.

PciConfigEntryIterator

Iterator providing a PciConfigEntry for all of the valid bus ranges on the system.

PciConfigRegions

Describes a set of regions of physical memory used to access the PCIe configuration space. A region is created for each entry in the MCFG. Given the segment group, bus, device number, and function of a PCIe device, the physical_address method on this will give you the physical address of the start of that device function’s configuration space (each function has 4096 bytes of configuration space in PCIe).