acpi

Module platform

Modules

interrupt

Structs

PlatformInfo

PlatformInfo allows the collection of some basic information about the platform from some of the fixed-size tables in a nice way. It requires access to the FADT and MADT. It is the easiest way to get information about the processors and interrupt controllers on a platform.

PmTimer

Information about the ACPI Power Management Timer (ACPI PM Timer).

Processor

ProcessorInfo

Enums

ProcessorState

Functions

wakeup_aps

Wakes up Application Processors (APs) using the Multiprocessor Wakeup Mailbox mechanism.