acpi::platform

Module interrupt

Structs

Apic

InterruptSourceOverride

Describes a difference in the mapping of an ISA interrupt to how it’s mapped in other interrupt models. For example, if a device is connected to ISA IRQ 0 and IOAPIC input 2, an override will appear mapping source 0 to GSI 2. Currently these will only be created for ISA interrupt sources.

IoApic

NmiLine

NmiSource

Describes a Global System Interrupt that should be enabled as non-maskable. Any source that is non-maskable can not be used by devices.

Enums

InterruptModel

LocalInterruptLine

Indicates which local interrupt line will be utilized by an external interrupt. Specifically, these lines directly correspond to their requisite LVT entries in a processor’s APIC.

NmiProcessor

Polarity

Polarity indicates what signal mode the interrupt line needs to be in to be considered ‘active’.

TriggerMode

Trigger mode of an interrupt, describing how the interrupt is triggered.