acpi::platform::interrupt

Struct Apic

pub struct Apic<'a, A>
where
    A: Allocator,
{
    pub local_apic_address: u64,
    pub io_apics: ManagedSlice<'a, IoApic, A>,
    pub local_apic_nmi_lines: ManagedSlice<'a, NmiLine, A>,
    pub interrupt_source_overrides: ManagedSlice<'a, InterruptSourceOverride, A>,
    pub nmi_sources: ManagedSlice<'a, NmiSource, A>,
    pub also_has_legacy_pics: bool,
}

Fields
local_apic_address: u64io_apics: ManagedSlice<'a, IoApic, A>local_apic_nmi_lines: ManagedSlice<'a, NmiLine, A>interrupt_source_overrides: ManagedSlice<'a, InterruptSourceOverride, A>nmi_sources: ManagedSlice<'a, NmiSource, A>also_has_legacy_pics: bool

If this field is set, you must remap and mask all the lines of the legacy PIC, even if you choose to use the APIC. Itโs recommended that you do this even if ACPI does not require you to.