Struct acpi::Acpi

pub struct Acpi {
    pub acpi_revision: u8,
    pub boot_processor: Option<Processor>,
    pub application_processors: Vec<Processor>,
    pub interrupt_model: Option<InterruptModel>,
    pub hpet: Option<HpetInfo>,
    pub power_profile: PowerProfile,
    pub dsdt: Option<AmlTable>,
    pub ssdts: Vec<AmlTable>,
    pub pci_config_regions: Option<PciConfigRegions>,
}

Fields
acpi_revision: u8boot_processor: Option<Processor>

The boot processor. Until you bring up any APs, this is the only processor running code.

application_processors: Vec<Processor>

Application processes. These are not brought up until you do so, and must be brought up in the order they appear in this list.

interrupt_model: Option<InterruptModel>

ACPI theoretically allows for more than one interrupt model to be supported by the same hardware. For simplicity and because hardware practically will only support one model, we just error in cases that the tables detail more than one.

hpet: Option<HpetInfo>power_profile: PowerProfiledsdt: Option<AmlTable>

Info about the DSDT, if we find it.

ssdts: Vec<AmlTable>

Info about any SSDTs, if there are any.

pci_config_regions: Option<PciConfigRegions>

Info about the PCI-E configuration memory regions, collected from the MCFG.