acpi::fadt

Struct Fadt

pub struct Fadt {

    pub sci_interrupt: u16,
    pub smi_cmd_port: u32,
    pub acpi_enable: u8,
    pub acpi_disable: u8,
    pub s4bios_req: u8,
    pub pstate_control: u8,
    pub gpe1_base: u8,
    pub c_state_control: u8,
    pub worst_c2_latency: u16,
    pub worst_c3_latency: u16,
    pub flush_size: u16,
    pub flush_stride: u16,
    pub duty_offset: u8,
    pub duty_width: u8,
    pub day_alarm: u8,
    pub month_alarm: u8,
    pub century: u8,
    pub iapc_boot_arch: IaPcBootArchFlags,
    pub flags: FixedFeatureFlags,
    pub reset_value: u8,
    pub arm_boot_arch: ArmBootArchFlags,
    /* private fields */

}

Represents the Fixed ACPI Description Table (FADT). This table contains various fixed hardware details, such as the addresses of the hardware register blocks. It also contains a pointer to the Differentiated Definition Block (DSDT).

In cases where the FADT contains both a 32-bit and 64-bit field for the same address, we should always prefer the 64-bit one. Only if itโs zero or the CPU will not allow us to access that address should the 32-bit one be used.

Fields
sci_interrupt: u16

On systems with an i8259 PIC, this is the vector the System Control Interrupt (SCI) is wired to. On other systems, this is the Global System Interrupt (GSI) number of the SCI.

The SCI should be treated as a sharable, level, active-low interrupt.

smi_cmd_port: u32

The system port address of the SMI Command Port. This port should only be accessed from the boot processor. A value of 0 indicates that System Management Mode is not supported.

Writing the value in acpi_enable to this port will transfer control of the ACPI hardware registers from the firmware to the OS. You must synchronously wait for the transfer to complete, indicated by the setting of SCI_EN.

Writing the value in acpi_disable will relinquish ownership of the hardware registers to the firmware. This should only be done if youโve previously acquired ownership. Before writing this value, the OS should mask all SCI interrupts and clear the SCI_EN bit.

Writing the value in s4bios_req requests that the firmware enter the S4 state through the S4BIOS feature. This is only supported if the S4BIOS_F flag in the FACS is set.

Writing the value in pstate_control yields control of the processor performance state to the OS. If this field is 0, this feature is not supported.

Writing the value in c_state_control tells the firmware that the OS supports _CST AML objects and notifications of C State changes.

acpi_enable: u8acpi_disable: u8s4bios_req: u8pstate_control: u8gpe1_base: u8c_state_control: u8worst_c2_latency: u16

The worst-case latency to enter and exit the C2 state, in microseconds. A value >100 indicates that the system does not support the C2 state.

worst_c3_latency: u16

The worst-case latency to enter and exit the C3 state, in microseconds. A value >1000 indicates that the system does not support the C3 state.

flush_size: u16flush_stride: u16duty_offset: u8duty_width: u8day_alarm: u8month_alarm: u8century: u8iapc_boot_arch: IaPcBootArchFlagsflags: FixedFeatureFlagsreset_value: u8arm_boot_arch: ArmBootArchFlags
Implementations

impl Fadt

pub fn validate(&self) -> Result<(), AcpiError>

pub fn facs_address(&self) -> Result<usize, AcpiError>

pub fn dsdt_address(&self) -> Result<usize, AcpiError>

pub fn power_profile(&self) -> PowerProfile

pub fn pm1a_event_block(&self) -> Result<GenericAddress, AcpiError>

pub fn pm1b_event_block(&self) -> Result<Option<GenericAddress>, AcpiError>

pub fn pm1a_control_block(&self) -> Result<GenericAddress, AcpiError>

pub fn pm1b_control_block(&self) -> Result<Option<GenericAddress>, AcpiError>

pub fn pm2_control_block(&self) -> Result<Option<GenericAddress>, AcpiError>

pub fn pm_timer_block(&self) -> Result<Option<GenericAddress>, AcpiError>

Attempts to parse the FADTโs PWM timer blocks, first returning the extended block, and falling back to parsing the legacy block into a GenericAddress.

pub fn gpe0_block(&self) -> Result<Option<GenericAddress>, AcpiError>

pub fn gpe1_block(&self) -> Result<Option<GenericAddress>, AcpiError>

pub fn reset_register(&self) -> Result<GenericAddress, AcpiError>

pub fn sleep_control_register( &self, ) -> Result<Option<GenericAddress>, AcpiError>

pub fn sleep_status_register(&self) -> Result<Option<GenericAddress>, AcpiError>