acpi::fadt

Struct FixedFeatureFlags

pub struct FixedFeatureFlags(/* private fields */);

Implementations

impl FixedFeatureFlags

pub fn supports_equivalent_to_wbinvd(&self) -> bool

If true, an equivalent to the x86 WBINVD instruction is supported. All caches will be flushed and invalidated upon completion of this instruction, and memory coherency is properly maintained. The cache SHALL only contain what OSPM references or allows to be cached.

pub fn wbinvd_flushes_all_caches(&self) -> bool

If true, WBINVD properly flushes all caches and memory coherency is maintained, but caches may not be invalidated.

pub fn all_procs_support_c1_power_state(&self) -> bool

If true, all processors implement the C1 power state.

pub fn c2_configured_for_mp_system(&self) -> bool

If true, the C2 power state is configured to work on a uniprocessor and multiprocessor system.

pub fn power_button_is_control_method(&self) -> bool

If true, the power button is handled as a control method device. If false, the power button is handled as a fixed-feature programming model.

pub fn sleep_button_is_control_method(&self) -> bool

If true, the sleep button is handled as a control method device. If false, the sleep button is handled as a fixed-feature programming model.

pub fn no_rtc_wake_in_fixed_register_space(&self) -> bool

If true, the RTC wake status is not supported in fixed register space.

pub fn rtc_wakes_system_from_s4(&self) -> bool

If true, the RTC alarm function can wake the system from an S4 sleep state.

pub fn pm_timer_is_32_bit(&self) -> bool

If true, indicates that the PM timer is a 32-bit value. If false, the PM timer is a 24-bit value and the remaining 8 bits are clear.

pub fn supports_docking(&self) -> bool

If true, the system supports docking.

pub fn supports_system_reset_via_fadt(&self) -> bool

If true, the system supports system reset via the reset_reg field of the FADT.

pub fn case_is_sealed(&self) -> bool

If true, the system supports no expansion capabilities and the case is sealed.

pub fn system_is_headless(&self) -> bool

If true, the system cannot detect the monitor or keyboard/mouse devices.

pub fn use_instr_after_write_to_slp_typx(&self) -> bool

If true, OSPM must use a processor instruction after writing to the SLP_TYPx register.

pub fn supports_pciexp_wake_in_pm1(&self) -> bool

If set, the platform supports the PCIEXP_WAKE_STS and PCIEXP_WAKE_EN bits in the PM1 status and enable registers.

pub fn use_pm_or_hpet_for_monotonically_decreasing_timers(&self) -> bool

If true, OSPM should use the ACPI power management timer or HPET for monotonically-decreasing timers.

pub fn rtc_sts_is_valid_after_wakeup_from_s4(&self) -> bool

If true, the contents of the RTC_STS register are valid after wakeup from S4.

pub fn ospm_may_leave_gpe_wake_events_armed_before_s5(&self) -> bool

If true, the platform supports OSPM leaving GPE wake events armed prior to an S5 transition.

pub fn lapics_must_use_cluster_model_for_logical_mode(&self) -> bool

If true, all LAPICs must be configured using the cluster destination model when delivering interrupts in logical mode.

pub fn local_xapics_must_use_physical_destination_mode(&self) -> bool

If true, all LXAPICs must be configured using physical destination mode.

pub fn system_is_hw_reduced_acpi(&self) -> bool

If true, this system is a hardware-reduced ACPI platform, and software methods are used for fixed-feature functions defined in chapter 4 of the ACPI specification.

pub fn no_benefit_to_s3(&self) -> bool

If true, the system can achieve equal or better power savings in an S0 power state, making an S3 transition useless.