acpi

Module madt

Structs

EntryHeader

GicInterruptTranslationServiceEntry

GicMsiFrameEntry

GicRedistributorEntry

GiccEntry

This field will appear for ARM processors that support ACPI and use the Generic Interrupt Controller. In the GICC interrupt model, each logical process has a Processor Device object in the namespace, and uses this structure to convey its GIC information.

GicdEntry

InterruptSourceOverrideEntry

IoApicEntry

IoSapicEntry

If this entry is present, the system has an I/O SAPIC, which must be used instead of the I/O APIC.

LocalApicAddressOverrideEntry

LocalApicEntry

LocalApicNmiEntry

LocalSapicEntry

LocalX2ApicEntry

Madt

Represents the MADT - this contains the MADT header fields. You can then iterate over a Madt to read each entry from it.

MadtEntryIter

MultiprocessorWakeupEntry

MultiprocessorWakeupMailbox

NmiSourceEntry

PlatformInterruptSourceEntry

X2ApicNmiEntry

Enums

MadtEntry

MadtError

MpProtectedModeWakeupCommand