acpi

Module sdt

Structs

ExtendedField

Represents a field which may or may not be present within an ACPI structure, depending on the version of ACPI that a system supports. If the field is not present, it is not safe to treat the data as initialised.

SdtHeader

All SDTs share the same header, and are length bytes long. The signature tells us which SDT this is.

Signature