acpi::sdt

Struct SdtHeader

#[repr(C, packed(1))]
pub struct SdtHeader {
    pub signature: Signature,
    pub length: u32,
    pub revision: u8,
    pub checksum: u8,
    pub oem_id: [u8; 6],
    pub oem_table_id: [u8; 8],
    pub oem_revision: u32,
    pub creator_id: u32,
    pub creator_revision: u32,
}

All SDTs share the same header, and are length bytes long. The signature tells us which SDT this is.

The ACPI Spec (Version 6.4) defines the following SDT signatures:

APIC - Multiple APIC Description Table (MADT)

BERT - Boot Error Record Table

BGRT - Boot Graphics Resource Table

CPEP - Corrected Platform Error Polling Table

DSDT - Differentiated System Description Table (DSDT)

ECDT - Embedded Controller Boot Resources Table

EINJ - Error Injection Table

ERST - Error Record Serialization Table

FACP - Fixed ACPI Description Table (FADT)

FACS - Firmware ACPI Control Structure

FPDT - Firmware Performance Data Table

GTDT - Generic Timer Description Table

HEST - Hardware Error Source Table

MSCT - Maximum System Characteristics Table

MPST - Memory Power StateTable

NFIT - NVDIMM Firmware Interface Table

OEMx - OEM Specific Information Tables

PCCT - Platform Communications Channel Table

PHAT - Platform Health Assessment Table

PMTT - Platform Memory Topology Table

PSDT - Persistent System Description Table

RASF - ACPI RAS Feature Table

RSDT - Root System Description Table

SBST - Smart Battery Specification Table

SDEV - Secure DEVices Table

SLIT - System Locality Distance Information Table

SRAT - System Resource Affinity Table

SSDT - Secondary System Description Table

XSDT - Extended System Description Table

Acpi reserves the following signatures and the specifications for them can be found here:

AEST - ARM Error Source Table

BDAT - BIOS Data ACPI Table

CDIT - Component Distance Information Table

CEDT - CXL Early Discovery Table

CRAT - Component Resource Attribute Table

CSRT - Core System Resource Table

DBGP - Debug Port Table

DBG2 - Debug Port Table 2 (note: ACPI 6.4 defines this as โDBPG2โ but this is incorrect)

DMAR - DMA Remapping Table

DRTM -Dynamic Root of Trust for Measurement Table

ETDT - Event Timer Description Table (obsolete, superseeded by HPET)

HPET - IA-PC High Precision Event Timer Table

IBFT - iSCSI Boot Firmware Table

IORT - I/O Remapping Table

IVRS - I/O Virtualization Reporting Structure

LPIT - Low Power Idle Table

MCFG - PCI Express Memory-mapped Configuration Space base address description table

MCHI - Management Controller Host Interface table

MPAM - ARM Memory Partitioning And Monitoring table

MSDM - Microsoft Data Management Table

PRMT - Platform Runtime Mechanism Table

RGRT - Regulatory Graphics Resource Table

SDEI - Software Delegated Exceptions Interface table

SLIC - Microsoft Software Licensing table

SPCR - Microsoft Serial Port Console Redirection table

SPMI - Server Platform Management Interface table

STAO - _STA Override table

SVKL - Storage Volume Key Data table (Intel TDX only)

TCPA - Trusted Computing Platform Alliance Capabilities Table

TPM2 - Trusted Platform Module 2 Table

UEFI - Unified Extensible Firmware Interface Specification table

WAET - Windows ACPI Emulated Devices Table

WDAT - Watch Dog Action Table

WDRT - Watchdog Resource Table

WPBT - Windows Platform Binary Table

WSMT - Windows Security Mitigations Table

XENV - Xen Project

Fields
signature: Signaturelength: u32revision: u8checksum: u8oem_id: [u8; 6]oem_table_id: [u8; 8]oem_revision: u32creator_id: u32creator_revision: u32
Implementations

impl SdtHeader

pub fn validate(&self, signature: Signature) -> AcpiResult<()>

Checks that:

The signature matches the one given.

The values of various fields in the header are allowed.

The checksum of the SDT is valid.

This assumes that the whole SDT is mapped.

pub fn oem_id(&self) -> &str

pub fn oem_table_id(&self) -> &str